requirements.txt.txt
flask
mysql-connector-python
psutil
pandas
scikit-learn
joblib
requests
python-nmap